requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
backports-abc==0.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
Cython==0.23
Django==1.10.6
docutils==0.13.1
enum34==1.1.6
future==0.17.1
futures==3.3.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-drive-api
httplib2==0.13.1
ipaddress==1.0.22
oauth2client==4.1.3
packaging==16.8
pipenv==11.8.0
psutil==5.0.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.2.0
pyowm==2.8.0
pyparsing==2.2.0
pyTelegramBotAPI
python-telegram-bot
requests==2.13.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
uritemplate==3.0.0
virtualenv==15.1.0
virtualenv-clone==0.3.0